+++ NEWS.md
@@ -1,4 +1,4 @@
-# waiter 0.2.2.9000
+# waiter 0.2.2
 
 - Remove `console.log`
 - Add `image` argument to the waiter, allows using a background